+++ src/com/nativebuild/util/ProgressBar.kt
@@ -46,10 +46,9 @@ private fun sizeOfDirectory(directory: File): Long {
 /**
  * Creates a progress bar based on several parameters.
  *
- * @param process The process of the file (to tell when the download is done)
  * @param filePath The path of the file that is is progressing
  * @param maxValue The value that the file should reach in bytes
- * @param barLength The length of the bar (default: 64)
+ * @param barLength The length of the bar. Maximum length is 128 (default: 64)
  * @param updateWait How long to wait between updates in milliseconds (default: 500)
  * @param reverse Reverses the progress bar (mainly used for when deleting something) (default: false)
  * @param unitMeasured What units the progress bar should be measured in. Valid values:
@@ -58,13 +57,13 @@ private fun sizeOfDirectory(directory: File): Long {
  * @return 0 if the progress bar successfully ran
  * @author Joshua Kent
  */
-fun progressBar(process: Process, filePath: String, maxValue: Long,
+fun progressBar(filePath: String, maxValue: Long,
                 barLength: Int = 64, updateWait: Long = 500, reverse: Boolean = false,
                 measuredIn: String? = null): Int {
 
     class ProgressBarException(message: String): Exception(message)
 
-    print("[" + ".".repeat(barLength) + "] 0%")
+    if (barLength > 100) { throw ProgressBarException("Maximum bar length (100) exceeded.") }
 
     val file = File(filePath)
 
@@ -78,58 +77,55 @@ fun progressBar(process: Process, filePath: String, maxValue: Long,
         MAXVALUE = maxValue / 1073741824
     } else { MAXVALUE = maxValue }
 
-    // main progress bar logic
-    var lastBarLength = 0
+    /* main progress bar logic */
+
+
+    // get file length depending on if it is a file or directory
+    var currentFileSize = if (file.exists()) {
+        if (file.isFile) { file.length() } else { sizeOfDirectory(file) }
+    } else { 0 }
+
+    // add units to end (ie. [134KB/5928KB])
+    when (measuredIn?.toUpperCase()) {
+        "KB" -> currentFileSize /= 1024
+        "MB" -> currentFileSize /= 1048576
+        "GB" -> currentFileSize /= 1073741824
+        !in arrayOf(null, "B") -> throw ProgressBarException(
+                "Unknown 'measuredIn' value (not 'B', 'KB', 'MB', 'GB', or null).")
+    }
+    val endUnits = if (measuredIn != null) {
+        " ($currentFileSize${measuredIn.toUpperCase()}/$MAXVALUE${measuredIn.toUpperCase()})"
+    } else { "" }
+
+    /* display progress */
+
+    val amountOfHashes = (currentFileSize.toFloat() / MAXVALUE.toFloat() * barLength).toInt()
+    val amountOfStops = barLength - amountOfHashes
+    val percentDone = (currentFileSize.toFloat() / MAXVALUE.toFloat() * 100).toInt()
+
     var barOutput = ""
-    var excess = ""
-    while (process.isAlive) {
-        // get file length depending on if it is a file or directory
-        var currentFileSize = if (file.exists()) {
-            if (file.isFile) { file.length() } else { sizeOfDirectory(file) }
-        } else { 0 }
-
-        // add units to end (ie. [134KB/5928KB])
-        var endUnits: String
-        when (measuredIn?.toUpperCase()) {
-            "KB" -> currentFileSize /= 1024
-            "MB" -> currentFileSize /= 1048576
-            "GB" -> currentFileSize /= 1073741824
-            !in arrayOf(null, "B") -> throw ProgressBarException(
-                    "Unknown 'measuredIn' value (not 'B', 'KB', 'MB', 'GB', or null).")
-        }
-        endUnits = if (measuredIn != null) {
-            " ($currentFileSize${measuredIn.toUpperCase()}/$MAXVALUE${measuredIn.toUpperCase()})"
-        } else { "" }
-
-        /* display progress */
-
-        var amountOfHashes = (currentFileSize.toFloat() / MAXVALUE.toFloat() * barLength).toInt()
-        var amountOfStops = barLength - amountOfHashes
-        var percentDone = (currentFileSize.toFloat() / MAXVALUE.toFloat() * 100).toInt()
-
-        lastBarLength = barOutput.length
-        if (currentFileSize >= MAXVALUE) {
-            if (!reverse) {
-                barOutput = "\r[" + "#".repeat(barLength) + "] 100%$endUnits"
-            } else {
-                barOutput = "\r[" + ".".repeat(barLength) + "] 0%$endUnits"
-            }
+    if (currentFileSize >= MAXVALUE) {
+        if (!reverse) {
+            barOutput = "\r[" + "#".repeat(barLength) + "] 100%$endUnits"
         } else {
-            if (!reverse) {
-                barOutput = "\r[" + "#".repeat(amountOfHashes) + ".".repeat(amountOfStops) + "] $percentDone%$endUnits"
-            } else {
-                barOutput = "\r[" + "#".repeat(amountOfStops) + ".".repeat(amountOfHashes) + "] ${100 - percentDone}%$endUnits"
-            }
+            barOutput = "\r[" + ".".repeat(barLength) + "] 0%$endUnits"
         }
-
-        // excess is to make sure that the entire stream of the current line gets wiped
-        excess = if (barOutput.length < lastBarLength) { " ".repeat(lastBarLength - barOutput.length) } else { "" }
-        print("\r$excess" + barOutput)
-
-        Thread.sleep(updateWait)
+    } else {
+        if (!reverse) {
+            barOutput = "\r[" + "#".repeat(amountOfHashes) + ".".repeat(amountOfStops) + "] $percentDone%$endUnits"
+        } else {
+            barOutput = "\r[" + "#".repeat(amountOfStops) + ".".repeat(amountOfHashes) + "] ${100 - percentDone}%$endUnits"
+        }
     }
 
-    print("\r" + " ".repeat(barOutput.length) + "\r")
-    //TODO: FIX ZIP FILES BEING UNDELETABLE (LOOK INTO java.io.ZipFile etc)
+
+    // remove excess based on largest amount the progress bar could be at 100% based on settings
+    var excessNum = barLength + 7
+    if (measuredIn != null) { excessNum += 7 + (measuredIn.length + MAXVALUE.toInt()) * 2 }
+
+
+    print(" ".repeat(excessNum) + "\r$barOutput\r")
+    Thread.sleep(updateWait)
+
     return 0
 }