+++ src/com/nativebuild/DATA.kt
@@ -16,16 +16,16 @@
 
 package com.nativebuild
 
-const val VERSION = "1.1-PRE12"
+const val VERSION = "1.1-PRE13"
 object LastUpdated {
     const val DATE = "02 July 2020"
-    const val TIME = "00:37 UTC+1"
+    const val TIME = "01:00 UTC+1"
     const val FULL = "$DATE $TIME"
 }
 const val FULLVERSION = "$VERSION at ${LastUpdated.FULL}"
 const val AUTHOR = "Joshua Kent"
 /** Set to `true` to skip anything currently building in `com.nativebuild.main()` */
-const val TESTING = true
+const val TESTING = "DEV" // set to DEV, BUILD or STABLE
 const val REPO = "https://github.com/joshua-kent/native-build"
 const val LICENSE = """                                 Apache License
                            Version 2.0, January 2004